# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "MeController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CurrentUserResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "MeController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCurrentUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CurrentUserResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "me"
        ]
      }
    },
...
    "/api/v1/patients/{patientId}": {
      "get": {
        "operationId": "PatientsController_detail",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PatientDetailResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "patients"
        ]
      },
      "patch": {
        "operationId": "PatientsController_update",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PatientResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "patients"
        ]
      }
    },
    "/api/v1/patients/{patientId}/consents": {
      "get": {
        "operationId": "ConsentsController_list",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ConsentResponseDto"
                          }
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "consents"
        ]
      }
    },
    "/api/v1/patients/{patientId}/consent-grants": {
      "post": {
        "operationId": "ConsentsController_grant",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsentGrantRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ConsentResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "consents"
        ]
      }
    },
    "/api/v1/patients/{patientId}/consent-withdrawals": {
      "post": {
        "operationId": "ConsentsController_withdraw",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsentWithdrawalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ConsentResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "consents"
        ]
      }
    },
    "/api/v1/patients/{patientId}/consents/{type}": {
      "put": {
        "operationId": "ConsentsController_setConsent",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetConsentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "consents"
        ]
      }
    },
...
    "/api/v1/practitioners/{practitionerId}/availability": {
      "get": {
        "operationId": "PractitionersController_availability",
        "parameters": [
          {
            "name": "practitionerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PractitionerAvailabilityResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "practitioners"
        ]
      }
    },
...
    "/api/v1/doctors/{doctorId}/availability": {
      "get": {
        "operationId": "DoctorsController_availability",
        "parameters": [
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "doctors"
        ]
      }
    },
...
    "/api/v1/audit/patients/{id}": {
      "get": {
        "operationId": "AuditController_patient",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "DermaHealth Backend API",
    "description": "Clinic workflow backend: identity, scheduling, clinical, workflow, EMR, CRM.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "UserMembershipResponseDto": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "clinicLocationId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "departmentId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "super_administrator",
              "patient",
              "doctor",
              "nurse",
              "receptionist",
              "lab_technician",
              "imaging_technician",
              "pharmacist",
              "care_coordinator",
              "customer_care_employee",
              "medical_administrator",
              "system_administrator",
              "clinical_process_designer"
            ]
          }
        },
        "required": [
          "organizationId",
          "clinicLocationId",
          "departmentId",
          "role"
        ]
      },
      "CurrentUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "object",
            "nullable": true
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true,
            "description": "Short-lived or transformed URL; null until T09 file service ships."
          },
          "status": {
            "type": "string",
            "enum": [
              "invited",
              "active",
              "locked",
              "disabled"
            ]
          },
          "activeOrganizationId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserMembershipResponseDto"
            }
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "avatarUrl",
          "status",
          "activeOrganizationId",
          "memberships",
          "version"
        ]
      },
...
      "UpdateCurrentUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "avatarFileId": {
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "version"
        ]
      },
...
      "ReferenceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "PatientResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "userId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "dob": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other",
              "unknown"
            ]
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "object",
            "nullable": true
          },
          "address": {
            "type": "object",
            "nullable": true
          },
          "bloodType": {
            "type": "string",
            "enum": [
              "A+",
              "A-",
              "B+",
              "B-",
              "AB+",
              "AB-",
              "O+",
              "O-",
              "unknown"
            ]
          },
          "primaryDoctor": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ReferenceResponseDto"
              }
            ]
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "code",
          "userId",
          "organizationId",
          "name",
          "dob",
          "gender",
          "phone",
          "email",
          "address",
          "bloodType",
          "primaryDoctor",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "ConsentSummaryItemDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "granted": {
            "type": "boolean"
          },
          "policyVersion": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "granted",
          "policyVersion"
        ]
      },
      "PatientDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "userId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "dob": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other",
              "unknown"
            ]
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "object",
            "nullable": true
          },
          "address": {
            "type": "object",
            "nullable": true
          },
          "bloodType": {
            "type": "string",
            "enum": [
              "A+",
              "A-",
              "B+",
              "B-",
              "AB+",
              "AB-",
              "O+",
              "O-",
              "unknown"
            ]
          },
          "primaryDoctor": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ReferenceResponseDto"
              }
            ]
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "activeAppointmentCount": {
            "type": "number"
          },
          "activeEncounterId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "activeCarePlanId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "consentSummary": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConsentSummaryItemDto"
            }
          }
        },
        "required": [
          "id",
          "code",
          "userId",
          "organizationId",
          "name",
          "dob",
          "gender",
          "phone",
          "email",
          "address",
          "bloodType",
          "primaryDoctor",
          "version",
          "createdAt",
          "updatedAt",
          "activeAppointmentCount",
          "activeEncounterId",
          "activeCarePlanId",
          "consentSummary"
        ]
      },
      "UpdatePatientRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "example": "1995-03-15"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other",
              "unknown"
            ]
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "object",
            "nullable": true
          },
          "address": {
            "type": "object",
            "nullable": true
          },
          "bloodType": {
            "type": "string",
            "enum": [
              "A+",
              "A-",
              "B+",
              "B-",
              "AB+",
              "AB-",
              "O+",
              "O-",
              "unknown"
            ]
          },
          "primaryDoctorId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "version"
        ]
      },
      "ConsentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "policyVersion": {
            "type": "string"
          },
          "granted": {
            "type": "boolean"
          },
          "grantedAt": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "withdrawnAt": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "patientId",
          "type",
          "policyVersion",
          "granted",
          "grantedAt",
          "withdrawnAt",
          "version"
        ]
      },
      "CreateConsentGrantRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "data_processing",
              "research_data_sharing",
              "telemedicine"
            ]
          },
          "policyVersion": {
            "type": "string"
          },
          "grantedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "type",
          "policyVersion"
        ]
      },
      "CreateConsentWithdrawalRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "data_processing",
              "research_data_sharing",
              "telemedicine"
            ]
          },
          "reason": {
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "type",
          "version"
        ]
      },
      "SetConsentRequest": {
        "type": "object",
        "properties": {}
      },
...
      "AvailabilitySlotResponseDto": {
        "type": "object",
        "properties": {
          "slotId": {
            "type": "string",
            "description": "Signed opaque reference required by booking/rescheduling."
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "remainingCapacity": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "slotId",
          "startsAt",
          "endsAt",
          "remainingCapacity"
        ]
      },
      "PractitionerAvailabilityResponseDto": {
        "type": "object",
        "properties": {
          "practitionerId": {
            "type": "string",
            "format": "uuid"
          },
          "clinicLocationId": {
            "type": "string",
            "format": "uuid"
          },
          "timezone": {
            "type": "object",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "slotDurationMinutes": {
            "type": "object",
            "nullable": true,
            "minimum": 5,
            "maximum": 240
          },
          "capacity": {
            "type": "object",
            "nullable": true,
            "minimum": 1
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AvailabilitySlotResponseDto"
            }
          }
        },
        "required": [
          "practitionerId",
          "clinicLocationId",
          "date",
          "slots"
        ]
      },
...
    }
  }
}